Add: models for service layer
