+++ backend/auth/internal/models/user.go
@@ -1,0 +1,12 @@
+package models
+
+type CreateUser struct {
+	Email        string
+	PasswordHash string
+}
+
+type User struct {
+	ID           int
+	Email        string
+	PasswordHash string
+}
